requirements.txt
ImageHash
numpy
opencv_python
piexif
Pillow
python-dotenv
scikit_learn
scikit-image
ttkbootstrap
rawpy
psycopg2-binary
google-genai
torch 
torchvision
transformers
face_recognition
git+https://github.com/openai/CLIP.git



